more profile and user abilities

+++ blabs.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, request, redirect, url_for, flash
+from flask import Flask, render_template, request, redirect, url_for, flash, session
 from bson.objectid import ObjectId
 import datetime
 
@@ -32,10 +32,21 @@ def new_blab():
 def blab_submit():
   now = datetime.datetime.now()
   # print(now.strftime("%m-%d-%Y %H:%M"))
-  blab = {
-    'text_content': request.form.get('text_content'),
-    'date': now,
-  }
+
+  if 'user_id' in session:
+    blab = {
+      'text_content': request.form.get('text_content'),
+      'date': now,
+      'user_id': session['user_id'],
+      'user_email': session['email'],
+    }
+  else:
+    blab = {
+      'text_content': request.form.get('text_content'),
+      'date': now,
+      'user_id': 'guest',
+      'user_email': 'guest'
+    }
   # WRITES TO THE blabs DB
   blabs.insert_one(blab)
 
@@ -54,26 +65,42 @@ def blab_update(blab_id):
   # The newly updated form data
   now = datetime.datetime.now()
   original_blab = blabs.find_one({'_id': ObjectId(blab_id)})
-  
-  updated_blab = {
-    'text_content': request.form.get('text_content'),
-    'date': original_blab['date'],
-    'updated_date': now
-  }
-  # Set that former blab from db to this updated one
-  blabs.update_one(
-    {'_id': ObjectId(blab_id)},
-    {'$set': updated_blab}
-  )
-  flash('Successfully edited blab!', 'info')
-  return redirect(url_for('blab_routes.all_blabs_index'))
+
+  if 'user_id' in session:
+    if original_blab['user_id'] == session['user_id']:
+      updated_blab = {
+        'text_content': request.form.get('text_content'),
+        'date': original_blab['date'],
+        'updated_date': now,
+        'user_id': session['user_id'],
+        'user_email': session['email'],
+      }
+      # Set that former blab from db to this updated one
+      blabs.update_one(
+        {'_id': ObjectId(blab_id)},
+        {'$set': updated_blab}
+      )
+      flash('Successfully edited blab!', 'info')
+      return redirect(url_for('blab_routes.all_blabs_index')) 
+  flash('You can only edit your own blabs!', 'danger')
+  return redirect(url_for('blab_routes.blab_show_one', blab_id=blab_id))
 
 # DELETE - a blab 
 # USING AN <a> TAG WITHOUT POST METHOD WORKS
 @blab_routes.route('/blabs/<blab_id>/delete')
-def donations_delete(blab_id):
-  blabs.delete_one({'_id': ObjectId(blab_id)})
-  flash('Successfully deleted a blab.', 'warning')
+def blabs_delete(blab_id):
+  blab = blabs.find_one({'_id': ObjectId(blab_id)})
+  if 'user_id' in session:
+    if blab['user_id'] == session['user_id']:
+      blabs.delete_one({'_id': ObjectId(blab_id)})
+      flash('Successfully deleted a blab.', 'warning')
+    else:
+      flash('You can only delete your own blabs.', 'danger')
+  elif 'user_id' not in session and blab['user_id'] == 'guest':
+    blabs.delete_one({'_id': ObjectId(blab_id)})
+    flash('Successfully deleted a blab.', 'warning')
+  else: 
+    flash('You can only delete your own blabs.', 'danger')
   return redirect(url_for('blab_routes.all_blabs_index'))
